fixed trap unit restore; rewriting methods

+++ Plugin.cs
@@ -25,9 +25,9 @@ namespace HullBreakerCompany
         public static bool compatibilityLQ = false;
 
         public static float BunkerEnemyScale;
-        public static float LandmineScale;
-        public static float SpikeTrapScale;
-        public static float TurretScale;
+        public static int LandmineScale;
+        public static int SpikeTrapScale;
+        public static int TurretScale;
         public static bool UseShortChatMessages;
         public static bool EnableEventMessages;
